Next attempt
Tried to get working path and assign it to pathv. Also tried messing with the readCommand function to actually updated commandLine. Created child process but didn't have it execute anything. Created an i int because we use it sometimes, and I hate recreating it. 

I have no idea if this is helpful at all, but hopefully it gets us closer to what we need. I've left some // TODO comments for what needs to be done (except under the child process, which hopefully is obvious, but if not, we need to execute the command found in the last array element).

Will catch up with you guys tomorrow.